% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_functions.r
\name{cdist}
\alias{cdist}
\title{Pairwise distances}
\usage{
cdist(X, Y, metric = c("euclidean", "minkowski", "manhattan"), p = 2)
}
\arguments{
\item{X, Y}{A matrix}

\item{metric}{The distance metric to use}

\item{p}{The power of the Minkowski distance}
}
\description{
Computes distance between each pair of the two collections of inputs. Returns a \code{matrix}.
}
\seealso{
\code{\link{rdist}}, \code{\link{pdist}}
}

