% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_metric.R
\name{is_metric}
\alias{is_metric}
\alias{is_distance_matrix}
\alias{triangle_inequality}
\title{Metric and triangle inequality}
\usage{
is_distance_matrix(mat, tolerance = .Machine$double.eps^0.5)

triangle_inequality(mat, tolerance = .Machine$double.eps^0.5)
}
\arguments{
\item{mat}{The matrix to evaluate}

\item{tolerance}{Differences smaller than tolerance are not reported.}
}
\description{
Does the distance matric come from a metric
}
\examples{
data <- matrix(rnorm(20), ncol = 2)
dm <- pdist(data)
is_distance_matrix(dm)
triangle_inequality(dm)

dm[1, 2] <- 1.1 * dm[1, 2]
is_distance_matrix(dm)
}
