% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_functions.r
\name{pdist}
\alias{pdist}
\title{Pairwise distances}
\usage{
pdist(X, metric = "euclidean")
}
\arguments{
\item{X}{A matrix}

\item{metric}{The distance metric to use}
}
\description{
Computes pairwise distance between observations.
}

