% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_functions.r, R/rdist-package.r
\docType{package}
\name{rdist}
\alias{cdist}
\alias{pdist}
\alias{rdist}
\alias{rdist-package}
\title{rdist: an R package for distances}
\usage{
rdist(X, metric = c("euclidean", "minkowski", "manhattan", "chebyshev",
  "maximum"), p = 2)

pdist(X, metric = c("euclidean", "minkowski", "manhattan", "chebyshev",
  "maximum"), p = 2)

cdist(X, Y, metric = c("euclidean", "minkowski", "manhattan", "chebyshev",
  "maximum"), p = 2)
}
\arguments{
\item{X, Y}{A matrix}

\item{metric}{The distance metric to use}

\item{p}{The power of the Minkowski distance}
}
\description{
\code{rdist} provide a common framework to calculate distances. There are three main functions: 
\itemize{
\item \code{rdist} computes the pairwise distances between observations in one matrix and returns a \code{dist} object,
\item \code{pdist} computes the pairwise distances between observations in one matrix and returns a \code{matrix}, and
\item \code{cdist} computes the distances between observations in two matrices and returns a \code{matrix}. 
}
}

