% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_functions.r
\name{cdist}
\alias{cdist}
\title{Pairwise distances}
\usage{
cdist(X, Y, metric = "euclidean")
}
\arguments{
\item{X, Y}{A matrix}

\item{metric}{The distance metric to use}
}
\description{
Computes distance between each pair of the two collections of inputs.
}

