% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farthest_point_sampling.R
\name{farthest_point_sampling}
\alias{farthest_point_sampling}
\title{Farthest point sampling}
\usage{
farthest_point_sampling(dist, k = nrow(dist), initial_point_index = 1L)
}
\arguments{
\item{dist}{Original distance matrix}

\item{k}{Number of points to sample}

\item{initial_point_index}{Index of p_1}
}
\description{
Farthest point sampling returns a reordering of the metric 
space P = {p_1, ..., p_k}, such that each p_i is the farthest 
point from the first i-1 points.
}
\examples{

# generate data
df <- matrix(runif(200), ncol = 2)
dist_mat <- rdist::pdist(df)
# farthest point sampling
fps <- farthest_point_sampling(dist_mat)
# have a look at the fps distance matrix
rdist(df[fps[1:5], ])
dist_mat[fps, fps][1:5, 1:5]
}
