% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_functions.r
\name{pdist}
\alias{pdist}
\title{Pairwise distances}
\usage{
pdist(X, metric = c("euclidean", "minkowski", "manhattan"), p = 2)
}
\arguments{
\item{X}{A matrix}

\item{metric}{The distance metric to use}

\item{p}{The power of the Minkowski distance}
}
\description{
Computes pairwise distance between observations.
}

